bokeh/requirements.txt
snowflake-connector-python
bokeh
pandas
posit-sdk
